Fix I/O deadlock issue
I found this in a larger project; Open3.popen3 requires either
threaded or select() based reading to avoid one full buffer on an
output stream blocking the process continuing to execute while youu
wait for input on another steram.

Open3.capture3() handles all of this for you if you don't mind not
streaming the read.